+++ cmp/grid/columns/Column.js
@@ -81,8 +81,10 @@ export class Column {
      * @param {boolean} [c.absSort] - true to enable absolute value sorting for this column.  If
      *      false (default) absolute value sorts will be ignored when cycling through the
      *     sortingOrder.
+     * @param {(string|Column~sortValueFn)} [c.sortValue] - alternate field name to reference or
+     *      function to call when producing a value for this column to be sorted by.
      * @param {Column~comparatorFn} [c.comparator] - function for comparing column values for
-     *     sorting
+     *     sorting.
      * @param {boolean} [c.resizable] - false to prevent user from drag-and-drop resizing.
      * @param {boolean} [c.movable] - false to prevent user from drag-and-drop re-ordering.
      * @param {boolean} [c.sortable] - false to prevent user from sorting on this column.
@@ -171,6 +173,7 @@ export class Column {
         rowHeight,
         absSort,
         sortingOrder,
+        sortValue,
         comparator,
         resizable,
         movable,
@@ -256,6 +259,7 @@ export class Column {
 
         this.absSort = withDefault(absSort, false);
         this.sortingOrder = withDefault(sortingOrder, Column.DEFAULT_SORTING_ORDER);
+        this.sortValue = sortValue;
         this.comparator = comparator;
 
         this.resizable = withDefault(resizable, true);
@@ -509,8 +513,16 @@ export class Column {
         }
 
         if (this.comparator === undefined) {
-            // Default comparator sorting to absValue-aware GridSorters in GridModel.sortBy[].
-            ret.comparator = this.defaultComparator;
+            // Use default comparator with appropriate inputs
+            ret.comparator = (valueA, valueB, agNodeA, agNodeB) => {
+                const recordA = agNodeA?.data,
+                    recordB = agNodeB?.data;
+
+                valueA = this.getSortValue(valueA, recordA),
+                valueB = this.getSortValue(valueB, recordB);
+
+                return this.defaultComparator(valueA, valueB);
+            };
         } else {
             // ...or process custom comparator with the Hoist-defined comparatorFn API.
             ret.comparator = (valueA, valueB, agNodeA, agNodeB) => {
@@ -532,6 +544,9 @@ export class Column {
                         agNodeB
                     };
 
+                valueA = this.getSortValue(valueA, recordA);
+                valueB = this.getSortValue(valueB, recordB);
+
                 return this.comparator(valueA, valueB, sortDir, abs, params);
             };
         }
@@ -548,6 +563,7 @@ export class Column {
     //--------------------
     // Implementation
     //--------------------
+    // Default comparator sorting to absValue-aware GridSorters in GridModel.sortBy[].
     defaultComparator = (v1, v2) => {
         const sortCfg = find(this.gridModel.sortBy, {colId: this.colId});
         return sortCfg ? sortCfg.comparator(v1, v2) : GridSorter.defaultComparator(v1, v2);
@@ -571,6 +587,15 @@ export class Column {
         if (pinned === true) return 'left';
         if (pinned === 'left' || pinned === 'right') return pinned;
         return null;
+    }
+
+    getSortValue(v, record) {
+        const {sortValue, gridModel} = this;
+        if (!sortValue) return v;
+
+        return isFunction(sortValue) ?
+            sortValue(v, {record, column: this, gridModel}) :
+            record?.data[sortValue] ?? v;
     }
 
 }
@@ -641,6 +666,13 @@ export function getAgHeaderClassFn(column) {
  * @param {*} value - cell data value (column + row).
  * @param {CellContext} context - additional data about the column, row and GridModel.
  * @return {*} - value for export.
+ */
+
+/**
+ * @callback Column~sortValueFn - function to return a value for sorting.
+ * @param {*} value - cell data value (column + row).
+ * @param {CellContext} context - additional data about the column, row and GridModel.
+ * @return {*} - value for sort.
  */
 
 /**